refactor: api 모듈에 있는 검증 클래스를 domain 모듈로 이동
